move some fixtures from test_sign_config to conftest for use in sign engine tests. update example_engine_fixture to use new config pattern. change create_users option to create_new_users
